Improved UART, added balance and display managers

+++ shared/algorithm/circular_buffer.h
@@ -8,16 +8,18 @@
 #ifndef CIRCULAR_BUFFER_H
 #define	CIRCULAR_BUFFER_H
 
+#include <stdint.h>
+
 typedef struct StructCircularBuffer {
-    int size;
-    int start;
+    uint16_t size;
+    uint16_t start;
     /* Tracking start and end of buffer would waste
      * one position. A full buffer would always have
      * to leave last position empty or otherwise
      * it would look empty. Instead this buffer uses
      * count to track if buffer is empty or full
      */
-    int count; // number of elements in buffer
+    uint16_t count; // number of elements in buffer
     /* Two ways to make buffer element type opaque
      * First is by using typedef for the element
      * pointer. Second is by using void pointer.
@@ -26,7 +28,7 @@ typedef struct StructCircularBuffer {
 } CircularBuffer;
 
 
-void make_cb(CircularBuffer *buffer, int size);
+void make_cb(CircularBuffer *buffer, uint16_t size);
 
 // Returns true is buffer is full.
 int cb_full(CircularBuffer *buffer);
